Playing with color in vga_buffer
Note: Right now, if you newline using a background color, the new row
will be blanked in that color. So even if you only write in 5 chars on
the next line in a new color, the color the row was blanked in will
still show up for the rest of the row.